+++ src/devcenter/HISTORY.rst
@@ -2,6 +2,11 @@
 
 Release History
 ===============
+0.2.0
+++++++
+* Update control plane and data plane APIs to v2023-04-01
+* Azure Deployment Environments GA release
+
 0.1.2
 ++++++
 * Remove hibernate from help example